Fix issue where if a new highlight is saved with tags, it would fail
This was because it would try to save the tags before saving the
highlight, and since the highlight isn't saved the tag wouldn't have the
user ID to save it to. So I needed to pass the user ID to the all_tags
method so it could be saved correctly.