Add validators for models and expanded model tests
Validators are set to ensure that relationships are correct. For
example, a highlight can only have a source and tags that the user of
the highlight owns. They can't link their highlight to someone else's
tag, for example. Also, there is a validator to ensure that all authors
linked to a source also are all owned by the same user.